Replacements feature; fixes #1393
Use replacements feature for mails and PDF contents
Use a trait for replacements
Fix failing tests on member birthdate
Remaining Texts old call
