+++ galette/lib/Galette/Entity/PdfContribution.php
@@ -75,115 +75,6 @@ abstract class PdfContribution extends PdfModel
     }
 
     /**
-     * Get patterns for a contribution
-     *
-     * @param boolean $legacy Whether to load legacy patterns
-     *
-     * @return array
-     */
-    protected function getContributionPatterns($legacy = true): array
-    {
-        $dynamic_patterns = $this->getDynamicPatterns('contrib');
-
-        $c_patterns = [
-            'contrib_label'     => [
-                'title'     => _T('Contribution label'),
-                'pattern'   => '/{CONTRIB_LABEL}/',
-            ],
-            'contrib_amount'    => [
-                'title'     => _T('Amount'),
-                'pattern'   => '/{CONTRIB_AMOUNT}/',
-            ],
-            'contrib_amount_letters' => [
-                'title'     => _T('Amount (in letters)'),
-                'pattern'   => '/{CONTRIB_AMOUNT_LETTERS}/',
-            ],
-            'contrib_date'      => [
-                'title'     => _T('Full date'),
-                'pattern'   => '/{CONTRIB_DATE}/',
-            ],
-            'contrib_year'      => [
-                'title'     => _T('Contribution year'),
-                'pattern'   => '/{CONTRIB_YEAR}/',
-            ],
-            'contrib_comment'   => [
-                'title'     => _T('Comment'),
-                'pattern'   => '/{CONTRIB_COMMENT}/',
-            ],
-            'contrib_bdate'     => [
-                'title'     => _T('Begin date'),
-                'pattern'   => '/{CONTRIB_BEGIN_DATE}/',
-            ],
-            'contrib_edate'     => [
-                'title'     => _T('End date'),
-                'pattern'   => '/{CONTRIB_END_DATE}/',
-            ],
-            'contrib_id'        => [
-                'title'     => _T('Contribution id'),
-                'pattern'   => '/{CONTRIB_ID}/',
-            ],
-            'contrib_payment'   => [
-                'title'     => _T('Payment type'),
-                'pattern'   => '/{CONTRIB_PAYMENT_TYPE}/'
-            ]
-        ];
-
-        if ($legacy === true) {
-            foreach ($c_patterns as $key => $pattern) {
-                $nkey = '_' . $key;
-                $pattern['pattern'] = str_replace(
-                    'CONTRIB_',
-                    'CONTRIBUTION_',
-                    $pattern['pattern']
-                );
-                $c_patterns[$nkey] = $pattern;
-            }
-        }
-
-        return $c_patterns + $dynamic_patterns;
-    }
-
-    /**
-     * Set contribution and proceed related replacements
-     *
-     * @param Contribution $contrib Contribution
-     *
-     * @return PdfModel
-     */
-    public function setContribution(Contribution $contrib): PdfContribution
-    {
-        global $login, $i18n;
-
-        $formatter = new NumberFormatter($i18n->getID(), NumberFormatter::SPELLOUT);
-
-        $c_replacements = [
-            'contrib_label'     => $contrib->type->libelle,
-            'contrib_amount'    => $contrib->amount,
-            'contrib_amount_letters' => $formatter->format($contrib->amount),
-            'contrib_date'      => $contrib->date,
-            'contrib_year'      => $contrib->raw_date->format('Y'),
-            'contrib_comment'   => $contrib->info,
-            'contrib_bdate'     => $contrib->begin_date,
-            'contrib_edate'     => $contrib->end_date,
-            'contrib_id'        => $contrib->id,
-            'contrib_payment'   => $contrib->spayment_type,
-        ];
-
-        foreach ($c_replacements as $key => $replacement) {
-            $nkey = '_' . $key;
-            $c_replacements[$nkey] = $replacement;
-        }
-        $this->setReplacements($c_replacements);
-
-        /** the list of all dynamic fields */
-        $fields = new \Galette\Repository\DynamicFieldsSet($this->zdb, $login);
-        $dynamic_fields = $fields->getList('contrib');
-        $this->setDynamicFields('contrib', $dynamic_fields, $contrib);
-
-        return $this;
-    }
-
-    /**
      * Build legend array
      *
      * @return array
