+++ tests/Galette/Entity/tests/units/Adherent.php
@@ -168,6 +168,9 @@ class Adherent extends atoum
      */
     private function dataAdherent(): array
     {
+        $bdate = new \DateTime(date('Y') . '-12-26');
+        //member is expected to be 82 years old
+        $bdate->sub(new \DateInterval('P82Y'));
         $data = [
             'nom_adh' => 'Durand',
             'prenom_adh' => 'René',
@@ -184,7 +187,7 @@ class Adherent extends atoum
             'sexe_adh' => 0,
             'prof_adh' => 'Chef de fabrication',
             'titre_adh' => null,
-            'ddn_adh' => '1937-12-26',
+            'ddn_adh' => $bdate->format('Y-m-d'),
             'lieu_naissance' => 'Gonzalez-sur-Meunier',
             'pseudo_adh' => 'ubertrand',
             'pays_adh' => 'Antarctique',
@@ -355,7 +358,7 @@ class Adherent extends atoum
             'sexe_adh' => 0,
             'prof_adh' => 'Chef de fabrication',
             'titre_adh' => null,
-            'ddn_adh' => '1934-06-08',
+            'ddn_adh' => 'NOT USED',
             'lieu_naissance' => 'Gonzalez-sur-Meunier',
             'pseudo_adh' => 'ubertrand',
             'cp_adh' => '39 069',
